Fixing System.IO.Pipelines.Tests that was broken for the netstandard configuration
