+++ src/System.IO.Pipelines/tests/PipeReaderWriterFacts.cs
@@ -16,7 +16,7 @@ using Xunit;
 
 namespace System.IO.Pipelines.Tests
 {
-    public class PipelineReaderWriterFacts : IDisposable
+    public partial class PipelineReaderWriterFacts : IDisposable
     {
         public PipelineReaderWriterFacts()
         {
@@ -150,42 +150,6 @@ namespace System.IO.Pipelines.Tests
             ReadOnlySequence<byte> buffer = result.Buffer;
 
             _pipe.Reader.Complete();
-        }
-
-        [Fact]
-        public async Task ResetAfterCompleteReaderAndWriterWithoutAdvancingClearsEverything()
-        {
-            _pipe.Writer.WriteEmpty(4094);
-            _pipe.Writer.WriteEmpty(4094);
-            await _pipe.Writer.FlushAsync();
-            ReadResult result = await _pipe.Reader.ReadAsync();
-            ReadOnlySequence<byte> buffer = result.Buffer;
-
-            SequenceMarshal.TryGetReadOnlySequenceSegment(
-                buffer,
-                out ReadOnlySequenceSegment<byte> start,
-                out int startIndex,
-                out ReadOnlySequenceSegment<byte> end,
-                out int endIndex);
-
-            var startSegment = (BufferSegment)start;
-            var endSegment = (BufferSegment)end;
-            Assert.NotNull(startSegment.MemoryOwner);
-            Assert.NotNull(endSegment.MemoryOwner);
-
-            _pipe.Reader.Complete();
-
-            // Nothing cleaned up
-            Assert.NotNull(startSegment.MemoryOwner);
-            Assert.NotNull(endSegment.MemoryOwner);
-
-            _pipe.Writer.Complete();
-
-            // Should be cleaned up now
-            Assert.Null(startSegment.MemoryOwner);
-            Assert.Null(endSegment.MemoryOwner);
-
-            _pipe.Reset();
         }
 
         [Fact]
